split and renamed events
