+++ services/director-v2/tests/unit/test_api_route_dynamic_scheduler.py
@@ -81,9 +81,7 @@ async def observed_service(
 
 @pytest.fixture
 def mock_scheduler_service_shutdown_tasks(mocker: MockerFixture) -> None:
-    module_base = (
-        "simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core._utils"
-    )
+    module_base = "simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core._events_utils"
     mocker.patch(f"{module_base}.service_push_outputs", autospec=True)
     mocker.patch(f"{module_base}.service_remove_containers", autospec=True)
     mocker.patch(
