+++ services/director-v2/tests/unit/test_modules_dynamic_sidecar_scheduler_task.py
@@ -23,12 +23,14 @@ from simcore_service_director_v2.modules.dynamic_sidecar.api_client._public impo
     DynamicSidecarClient,
 )
 from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core import (
+    _events_utils,
     _observer,
-    _utils,
+)
+from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core._abc import (
+    DynamicSchedulerEvent,
 )
 from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core._events import (
     REGISTERED_EVENTS,
-    DynamicSchedulerEvent,
 )
 from simcore_service_director_v2.modules.dynamic_sidecar.scheduler._task import (
     DynamicSidecarsScheduler,
@@ -206,7 +208,7 @@ def mocked_dynamic_scheduler_events(
 
 @pytest.fixture
 def mock_remove_calls(mocker: MockerFixture) -> None:
-    mocker.patch.object(_utils, "remove_volumes_from_node")
+    mocker.patch.object(_events_utils, "remove_volumes_from_node")
 
 
 async def test_skip_observation_cycle_after_error(
